+++ src/index.ts
@@ -1,14 +1,11 @@
 class Greeter {
   public greeting: string;
   public constructor(message: string) {
-      this.greeting = message;
+    this.greeting = message;
   }
   public greet() {
-    return "Hello, "+this.greeting;
+    return "Hello, " + this.greeting;
   }
 }
 
-
-
-export default Greeter
-
+export default Greeter;