+++ test/greet.spec.ts
@@ -1,0 +1,9 @@
+import Greeter from '@/greet';
+
+describe('src/greet.ts', () => {
+  it('name param test', () => {
+    const greeter = new Greeter('world');
+
+    expect(greeter.greet()).toBe('Hello, from world');
+  });
+});