+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ### Features
 
-* **build:** 新增 ESLint 代码校验能力 ([ccc5638](https://github.com/SUN-LG/typescript-starter/commit/ccc5638d97c20fd9a83e0b5808b623e893cde2dc))
+* **build:** 新增 Prettier 自动格式化能力 ([557a3b8](https://github.com/SUN-LG/typescript-starter/commit/557a3b878398ff47d7712e640eec196f5cd3f563))
+* **build:** 新增 ESLint 代码校验能力 ([1d7477b](https://github.com/SUN-LG/typescript-starter/commit/1d7477b8d07f04292f9c8038e3d02661a1805192))
 * **build:** 新增 Git Commit Message 提交规范及校验 ([d2ac9ac](https://github.com/SUN-LG/typescript-starter/commit/d2ac9acbfc99a3b7d06974f9b0ce744245fa8cbb))
 * **build:** 新增 Typescript 编译能力 ([0b0f05b](https://github.com/SUN-LG/typescript-starter/commit/0b0f05bfdc88ac806accda64099a831b58e0e90b))
 * 项目初始化 ([8b28b61](https://github.com/SUN-LG/typescript-starter/commit/8b28b61228e5f5e010c7a3388c9807bdcd2c97d4))