+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-# 1.0.0 (2020-08-16)
+# 1.0.0 (2020-08-25)
 
 
 ### Features
 
 * **build:** 新增 ESLint 代码校验能力 ([1d7477b](https://github.com/SUN-LG/typescript-starter/commit/1d7477b8d07f04292f9c8038e3d02661a1805192))
 * **build:** 新增 Git Commit Message 提交规范及校验 ([d2ac9ac](https://github.com/SUN-LG/typescript-starter/commit/d2ac9acbfc99a3b7d06974f9b0ce744245fa8cbb))
-* **build:** 新增 Lint Staged 上传校验功能 ([cb3f71b](https://github.com/SUN-LG/typescript-starter/commit/cb3f71bc15572b6c063ebe606d1ed557e0c8e063))
+* **build:** 新增 Jest 单元测试功能 ([1a808c8](https://github.com/SUN-LG/typescript-starter/commit/1a808c891b3ebcb098012a9e761dc08dc930f286))
+* **build:** 新增 Lint Staged 上传校验功能 ([282e574](https://github.com/SUN-LG/typescript-starter/commit/282e574235141694cd8d41e4d3b900814e19614d))
 * **build:** 新增 Prettier 自动格式化能力 ([e43849f](https://github.com/SUN-LG/typescript-starter/commit/e43849f4239fbbec3f430d6f86db62e55856e4f6))
 * **build:** 新增 Typescript 编译能力 ([0b0f05b](https://github.com/SUN-LG/typescript-starter/commit/0b0f05bfdc88ac806accda64099a831b58e0e90b))
 * 项目初始化 ([8b28b61](https://github.com/SUN-LG/typescript-starter/commit/8b28b61228e5f5e010c7a3388c9807bdcd2c97d4))